requirements.txt
modbus-tk
pyserial
paho-mqtt
configparser
pytz
requests
schedule